Controle da dados do susuario logado

+++ src/main/java/dao/DAOUsuarioRepository.java
@@ -3,7 +3,6 @@ package dao;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,17 +19,19 @@ public class DAOUsuarioRepository {
 		connection = SingleConnectionBanco.getConnection();
 	}
 	
-	public ModelLogin gravarUsuario(ModelLogin objeto) throws Exception {
+	public ModelLogin gravarUsuario(ModelLogin objeto, Long userLogado) throws Exception {
 		
 		if (objeto.isNovo()) {/*Grava um novo*/
 		
-		String sql = "INSERT INTO model_login(login, senha, nome, email)  VALUES (?, ?, ?, ?);";
+		String sql = "INSERT INTO model_login(login, senha, nome, email, usuario_id )  VALUES (?, ?, ?, ?, ?);";
 		PreparedStatement preparedSql = connection.prepareStatement(sql);
 		
 		preparedSql.setString(1, objeto.getLogin());
 		preparedSql.setString(2, objeto.getSenha());
 		preparedSql.setString(3, objeto.getNome());
 		preparedSql.setString(4, objeto.getEmail());
+		preparedSql.setLong(5, userLogado);
+		
 		
 		preparedSql.execute();
 		
@@ -46,6 +47,7 @@ public class DAOUsuarioRepository {
 			prepareSql.setString(3, objeto.getNome());
 			prepareSql.setString(4, objeto.getEmail());
 			
+			
 			prepareSql.executeUpdate();
 			
 			connection.commit();
@@ -53,58 +55,70 @@ public class DAOUsuarioRepository {
 		}
 		
 		
-		return this.consultaUsuario(objeto.getLogin());
+		return this.consultaUsuario(objeto.getLogin(), userLogado);
 	}
 	
-public List<ModelLogin> consultaUsuarioList() throws Exception{
+	
+	
+	public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception {
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "select * from model_login ";
+		String sql = "select * from model_login where useradmin is false and usuario_id = " + userLogado;
 		PreparedStatement statement = connection.prepareStatement(sql);
 		
-		
 		ResultSet resultado = statement.executeQuery();
-		while(resultado.next()) { /*Percorrer as linhas de resultado do SQL*/
+		
+		while (resultado.next()) { /*percorrer as linhas de resultado do SQL*/
+			
 			ModelLogin modelLogin = new ModelLogin();
+			
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setId(resultado.getLong("id"));
 			modelLogin.setLogin(resultado.getString("login"));
 			modelLogin.setNome(resultado.getString("nome"));
 			//modelLogin.setSenha(resultado.getString("senha"));
+			
+			
 			retorno.add(modelLogin);
 		}
 		
-		return retorno;
 		
+		return retorno;
 	}
 	
-
-	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception{
+	
+	public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception {
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "select * from model_login where upper(nome) like upper (?)";
+		String sql = "select * from model_login  where upper(nome) like upper(?) and useradmin is false and usuario_id = ?";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setString(1, "%" + nome + "%");
+		statement.setLong(2, userLogado);
 		
 		ResultSet resultado = statement.executeQuery();
-		while(resultado.next()) { /*Percorrer as linhas de resultado do SQL*/
+		
+		while (resultado.next()) { /*percorrer as linhas de resultado do SQL*/
+			
 			ModelLogin modelLogin = new ModelLogin();
+			
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setId(resultado.getLong("id"));
 			modelLogin.setLogin(resultado.getString("login"));
 			modelLogin.setNome(resultado.getString("nome"));
 			//modelLogin.setSenha(resultado.getString("senha"));
+			
+			
 			retorno.add(modelLogin);
 		}
 		
-		return retorno;
 		
+		return retorno;
 	}
 	
 	
-	public ModelLogin consultaUsuario(String login) throws Exception  {
+	public ModelLogin consultaUsuarioLogado(String login) throws Exception  {
 		
 		ModelLogin modelLogin = new ModelLogin();
 		
@@ -121,6 +135,7 @@ public List<ModelLogin> consultaUsuarioList() throws Exception{
 			modelLogin.setLogin(resutlado.getString("login"));
 			modelLogin.setSenha(resutlado.getString("senha"));
 			modelLogin.setNome(resutlado.getString("nome"));
+			
 		}
 		
 		
@@ -129,14 +144,15 @@ public List<ModelLogin> consultaUsuarioList() throws Exception{
 	}
 	
 	
-public ModelLogin consultaUsuarioID(String id) throws Exception  {
+	
+	
+	public ModelLogin consultaUsuario(String login) throws Exception  {
 		
 		ModelLogin modelLogin = new ModelLogin();
 		
-		String sql = "select * from model_login where id = ?";
+		String sql = "select * from model_login where upper(login) = upper('"+login+"') and useradmin is false ";
 		
 		PreparedStatement statement = connection.prepareStatement(sql);
-		statement.setLong(1, Long.parseLong(id));
 		
 		ResultSet resutlado =  statement.executeQuery();
 		
@@ -147,12 +163,69 @@ public ModelLogin consultaUsuarioID(String id) throws Exception  {
 			modelLogin.setLogin(resutlado.getString("login"));
 			modelLogin.setSenha(resutlado.getString("senha"));
 			modelLogin.setNome(resutlado.getString("nome"));
+			
 		}
 		
 		
 		return modelLogin;
 		
 	}
+	
+	
+	
+	public ModelLogin consultaUsuario(String login, Long userLogado) throws Exception  {
+		
+		ModelLogin modelLogin = new ModelLogin();
+		
+		String sql = "select * from model_login where upper(login) = upper('"+login+"') and useradmin is false and usuario_id = " + userLogado;
+		
+		PreparedStatement statement = connection.prepareStatement(sql);
+		
+		ResultSet resutlado =  statement.executeQuery();
+		
+		while (resutlado.next()) /*Se tem resultado*/ {
+			
+			modelLogin.setId(resutlado.getLong("id"));
+			modelLogin.setEmail(resutlado.getString("email"));
+			modelLogin.setLogin(resutlado.getString("login"));
+			modelLogin.setSenha(resutlado.getString("senha"));
+			modelLogin.setNome(resutlado.getString("nome"));
+			
+		}
+		
+		
+		return modelLogin;
+		
+	}
+	
+	
+	public ModelLogin consultaUsuarioID(String id, Long userLogado) throws Exception  {
+		
+		ModelLogin modelLogin = new ModelLogin();
+		
+		String sql = "select * from model_login where id = ? and useradmin is false and usuario_id = ?";
+		
+		PreparedStatement statement = connection.prepareStatement(sql);
+		statement.setLong(1, Long.parseLong(id));
+		statement.setLong(2, userLogado);
+		
+		ResultSet resutlado =  statement.executeQuery();
+		
+		while (resutlado.next()) /*Se tem resultado*/ {
+			
+			modelLogin.setId(resutlado.getLong("id"));
+			modelLogin.setEmail(resutlado.getString("email"));
+			modelLogin.setLogin(resutlado.getString("login"));
+			modelLogin.setSenha(resutlado.getString("senha"));
+			modelLogin.setNome(resutlado.getString("nome"));
+			
+		}
+		
+		
+		return modelLogin;
+		
+	}
+	
 	
 	
 	public boolean validarLogin(String login) throws Exception {
@@ -167,14 +240,18 @@ public ModelLogin consultaUsuarioID(String id) throws Exception  {
 		
 	}
 	
-	public void deleteUser(String idUser) throws Exception{
-		String sql = "DELETE FROM model_login WHERE id = ?;";
+	
+	public void deletarUser(String idUser) throws Exception {
+		String sql = "DELETE FROM model_login WHERE id = ? and useradmin is false;";
+		
 		PreparedStatement prepareSql = connection.prepareStatement(sql);
 		
 		prepareSql.setLong(1, Long.parseLong(idUser));
+		
 		prepareSql.executeUpdate();
 		
 		connection.commit();
+		
 	}
 
 }
